Various tweaks in the documentation for linter complaints
I'd hesitate to always build this documentation with the linter switched
on, since there are some things it complains about that wouldn't make
sense for the Ecmarkup documentation. But there are some things it
complains about where it makes sense for the Ecmarkup docs to set a good
example.
pick