Fix emu-format --check with multiple files (#607)
Previously, running emu-format --check with multiple files would only
check the first file. Fix the CLI arguments processing logic so that the
only special case for a list of one file is when it needs to be printed
to stdout. Move checking into the loop through the file list, as is done
with writing.

Closes: #606